backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.21
pymysql==1.1.0
cryptography==41.0.4
gunicorn==21.2.0
python-dotenv==1.0.0
